remove inspect button + move files out of wizard folder
