+++ x-pack/plugins/observability_onboarding/public/components/app/custom_logs/api_key_banner.tsx
@@ -18,7 +18,7 @@ import { IHttpFetchError, ResponseErrorBody } from '@kbn/core-http-browser';
 import { i18n } from '@kbn/i18n';
 import { FETCH_STATUS } from '@kbn/observability-shared-plugin/public';
 import React from 'react';
-import { APIReturnType } from '../../../../services/rest/create_call_api';
+import { APIReturnType } from '../../../services/rest/create_call_api';
 
 type ApiKeyPayload =
   APIReturnType<'POST /internal/observability_onboarding/logs/flow'>;
