+++ x-pack/plugins/observability_onboarding/public/components/app/custom_logs/install_elastic_agent.tsx
@@ -7,10 +7,7 @@
 
 import {
   EuiButton,
-  EuiButtonEmpty,
   EuiCallOut,
-  EuiFlexGroup,
-  EuiFlexItem,
   EuiHorizontalRule,
   EuiSpacer,
   EuiText,
@@ -18,28 +15,28 @@ import {
 import { i18n } from '@kbn/i18n';
 import { useKibana } from '@kbn/kibana-react-plugin/public';
 import { default as React, useCallback, useEffect, useState } from 'react';
-import { ObservabilityOnboardingPluginSetupDeps } from '../../../../plugin';
+import { ObservabilityOnboardingPluginSetupDeps } from '../../../plugin';
 import { useWizard } from '.';
-import { FETCH_STATUS, useFetcher } from '../../../../hooks/use_fetcher';
+import { FETCH_STATUS, useFetcher } from '../../../hooks/use_fetcher';
 import {
   ElasticAgentPlatform,
   getElasticAgentSetupCommand,
-} from '../../../shared/get_elastic_agent_setup_command';
+} from '../../shared/get_elastic_agent_setup_command';
 import {
   InstallElasticAgentSteps,
   ProgressStepId,
   EuiStepStatus,
-} from '../../../shared/install_elastic_agent_steps';
+} from '../../shared/install_elastic_agent_steps';
 import {
   StepPanel,
   StepPanelContent,
   StepPanelFooter,
-} from '../../../shared/step_panel';
+} from '../../shared/step_panel';
 import { ApiKeyBanner } from './api_key_banner';
 import { BackButton } from './back_button';
-import { getDiscoverNavigationParams } from '../../utils';
-import { WindowsInstallStep } from '../../../shared/windows_install_step';
-import { TroubleshootingLink } from '../../../shared/troubleshooting_link';
+import { getDiscoverNavigationParams } from '../utils';
+import { WindowsInstallStep } from '../../shared/windows_install_step';
+import { TroubleshootingLink } from '../../shared/troubleshooting_link';
 
 export function InstallElasticAgent() {
   const {
@@ -47,14 +44,11 @@ export function InstallElasticAgent() {
       discover: { locator },
     },
   } = useKibana<ObservabilityOnboardingPluginSetupDeps>();
-  const { goBack, goToStep, getState, setState } = useWizard();
+  const { goBack, getState, setState } = useWizard();
   const wizardState = getState();
   const [elasticAgentPlatform, setElasticAgentPlatform] =
     useState<ElasticAgentPlatform>('linux-tar');
 
-  function onInspect() {
-    goToStep('inspect');
-  }
   async function onContinue() {
     await locator?.navigate(
       getDiscoverNavigationParams([wizardState.datasetName])
@@ -235,33 +229,18 @@ export function InstallElasticAgent() {
         <StepPanelFooter
           items={[
             <BackButton onBack={goBack} />,
-            <EuiFlexGroup justifyContent="flexEnd" alignItems="center">
-              <EuiFlexItem grow={false}>
-                <EuiButtonEmpty
-                  data-test-subj="observabilityOnboardingInstallElasticAgentInspectButton"
-                  onClick={onInspect}
-                >
-                  {i18n.translate(
-                    'xpack.observability_onboarding.steps.inspect',
-                    { defaultMessage: 'Inspect' }
-                  )}
-                </EuiButtonEmpty>
-              </EuiFlexItem>
-              <EuiFlexItem grow={false}>
-                <EuiButton
-                  data-test-subj="obltOnboardingExploreLogs"
-                  color="success"
-                  fill
-                  iconType="magnifyWithPlus"
-                  onClick={onContinue}
-                >
-                  {i18n.translate(
-                    'xpack.observability_onboarding.steps.exploreLogs',
-                    { defaultMessage: 'Explore logs' }
-                  )}
-                </EuiButton>
-              </EuiFlexItem>
-            </EuiFlexGroup>,
+            <EuiButton
+              data-test-subj="obltOnboardingExploreLogs"
+              color="success"
+              fill
+              iconType="magnifyWithPlus"
+              onClick={onContinue}
+            >
+              {i18n.translate(
+                'xpack.observability_onboarding.steps.exploreLogs',
+                { defaultMessage: 'Explore logs' }
+              )}
+            </EuiButton>,
           ]}
         />
       }
