+++ x-pack/plugins/observability_onboarding/public/components/app/custom_logs/select_logs.tsx
@@ -25,9 +25,9 @@ import {
   StepPanel,
   StepPanelContent,
   StepPanelFooter,
-} from '../../../shared/step_panel';
+} from '../../shared/step_panel';
 import { useWizard } from '.';
-import { useKibanaNavigation } from '../../../../hooks/use_kibana_navigation';
+import { useKibanaNavigation } from '../../../hooks/use_kibana_navigation';
 
 export function SelectLogs() {
   const { navigateToKibanaUrl, navigateToAppUrl } = useKibanaNavigation();
