+++ x-pack/plugins/observability_onboarding/public/components/app/system_logs/install_elastic_agent.tsx
@@ -35,7 +35,7 @@ import {
   StepPanelContent,
   StepPanelFooter,
 } from '../../shared/step_panel';
-import { ApiKeyBanner } from '../custom_logs/wizard/api_key_banner';
+import { ApiKeyBanner } from '../custom_logs/api_key_banner';
 import { getDiscoverNavigationParams } from '../utils';
 import { WindowsInstallStep } from '../../shared/windows_install_step';
 import { SystemIntegrationBanner } from './system_integration_banner';
